cleanup leg routes
