+++ routes/legs.ts
@@ -14,171 +14,18 @@
  * limitations under the License.
  */
 import { Request, Response } from 'express';
-import { formatTime, ranking, parseTime } from '@rasifix/orienteering-utils';
-import { EventLoader, Category, Runner } from '../types/index.ts';
-import { RankingRunner } from '@rasifix/orienteering-utils/lib/utils/ranking';
+import { EventLoader } from '../types/index.ts';
+import { buildLegs } from '../services/leg-builder.ts';
 
 export default function(loader: EventLoader) {
   return (req: Request, res: Response) => {
     const id = req.params.id;
     loader(id, (event) => {
-      const legs = defineLegs(event.categories);
-    
+      const legs = buildLegs(event.categories);
       res.json(legs);
     }, (error) => {
       res.status(error.statusCode);
       res.json(error);
     });
   };
-}
-
-interface Leg {
-  id: string;
-  from: string;
-  to: string;
-  categories: { [name: string]: boolean };
-  runners: LegRunner[];
-  errorFrequency: number;
-}
-
-interface LegRunner {
-  id: string;
-  fullName: string;
-  yearOfBirth?: number;
-  city?: string;
-  club?: string;
-  split: string;
-  splitBehind?: string;
-  splitRank?: number;
-  category: string;
-  timeLoss?: string;
-} 
-
-function defineLegs(categories: Category[]) {
-  // helper function to create ranking entry for runner
-  const createRankingEntry = (runner: Runner, category: string, splitTime: number) => {
-    return {
-      id: runner.id,
-      fullName: runner.fullName,
-      yearOfBirth: runner.yearOfBirth,
-      city: runner.city,
-      club: runner.club,
-      split: formatTime(splitTime),
-      category: category
-    };
-  };
-  
-  const legs: { [key: string]: Leg } = {};
-  let lastSplit = null;
-  categories.forEach((category) => {        
-    category.runners.forEach((runner) => {
-      let lastTime: string | null = null;
-      let lastControl = 'St';
-      runner.splits!.forEach((split) => {
-        const control = split.code;
-        const time = split.time;
-        const code = lastControl + '-' + control;
-        if (!legs[code]) {
-          legs[code] = {
-            id: code,
-            from: lastControl,
-            to: control,
-            categories: {},
-            errorFrequency: 0,
-            runners: []
-          };
-        }
-        if (isValid(time) && (lastTime == null || isValid(lastTime))) {
-          const splitTime = lastTime !== null ? (parseTime(time) ?? 0) - (parseTime(lastTime) ?? 0) : (parseTime(time) ?? 0);
-          legs[code].runners.push(createRankingEntry(runner, category.name, splitTime));
-          legs[code].categories[category.name] = true;
-          lastSplit = split;
-        }
-      
-        lastControl = control;
-        lastTime = time;
-      });
-    });
-  });
-  
-  // convert legs hash into array
-  const result: Leg[] = [];
-  Object.keys(legs).forEach((code) => {
-    const leg = legs[code];
-    leg.runners.sort((s1: LegRunner, s2: LegRunner) => {
-      return (parseTime(s1.split) ?? 0) - (parseTime(s2.split) ?? 0);
-    });
-    leg.errorFrequency = 0;
-    if (leg.runners.length > 0) {
-      result.push(leg);
-    }
-  });
-  
-  const runners: { [id: string]: RankingRunner } = {};
-  categories.forEach((c) => {
-    const runnersFormatted = c.runners.map(r => ({
-      ...r,
-      id: String(r.id),
-      category: c.name,
-      startTime: r.starttime || '',
-      yearOfBirth: r.yearOfBirth?.toString(),
-      splits: r.splits || []
-    }));
-    const categoryParsed = ranking.parseRanking(runnersFormatted);
-    categoryParsed.runners.forEach((runner) => {
-      runners[runner.id] = runner;
-    });
-  });
-  
-  result.forEach((leg) => {
-    let timeLosses = 0;
-    const fastest = leg.runners.length > 0 ? (parseTime(leg.runners[0].split) ?? 0) : 0;
-    
-    leg.runners.forEach((runner, idx: number) => {
-      const r = runners[runner.id];
-      const s = r.splits.find((split) => leg.id === split.legCode);
-      if (s && s.timeLoss) {
-        timeLosses += 1;
-        runner.timeLoss = formatTime(s.timeLoss);
-      }
-      
-      if (idx > 0) {
-        runner.splitBehind = '+' + formatTime((parseTime(runner.split) ?? 0) - fastest);
-      }
-      
-      if (idx === 0) {
-        runner.splitRank = 1;
-      } else {
-        const prev = leg.runners[idx - 1];
-        if (prev.split === runner.split) {
-          runner.splitRank = prev.splitRank;
-        } else {
-          runner.splitRank = idx + 1;
-        }
-      }
-    });
-    
-    if (leg.runners.length > 0) {
-      leg.errorFrequency = Math.round(100 * timeLosses / leg.runners.length);
-    }
-  });
-      
-  result.sort((l1, l2) => {
-    return l2.errorFrequency - l1.errorFrequency;
-  });
-  
-  return result.map((leg) => {
-    return {
-      id: leg.id,
-      from: leg.from,
-      to: leg.to,
-      categories: Object.keys(leg.categories),
-      runners: leg.runners.length,
-      errorFrequency: leg.errorFrequency
-    };
-  });
-}
-
-function isValid(value: string): boolean {
-  return value !== '-' && value !== 's' && parseTime(value) !== null;
 }
